requirements.txt
fastapi==0.115.0
pydantic==2.8.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
